新特性: 适配新功能(v1.8.0)
- 新增 发送Protobuf数据包
- 新增 群文件转永久
- 修改 引用消息片段实体参数类型定义

+++ src/main/java/ink/on/central/bot/entity/request/msg/part/LLOBMsgReply.java
@@ -22,7 +22,7 @@ public class LLOBMsgReply extends LLOBMsgPart {
    *
    * @param id 消息id
    */
-  public LLOBMsgReply(@NotNull Integer id) {
+  public LLOBMsgReply(@NotNull Long id) {
     super("reply");
     this.data = new LLOBMsgReplyData().setId(id);
   }
@@ -37,7 +37,7 @@ public class LLOBMsgReply extends LLOBMsgPart {
   @EqualsAndHashCode(callSuper = true)
   public static class LLOBMsgReplyData extends LLOBMsgBase {
     /** 回复的消息id */
-    private Integer id;
+    private Long id;
   }
 
 }
